LayoutTests/svg/dynamic-updates/SVGFECompositeElement-svgdom-in2-prop-expected.txt

CONSOLE MESSAGE: line 116: TypeError: 'undefined' is not an object (evaluating 'overComposite1.in2.baseVal = "off1"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in2' property of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL overComposite1.in2.baseVal should be SourceGraphic. Threw exception TypeError: 'undefined' is not an object (evaluating 'overComposite1.in2.baseVal')
